chore(core): cleanup
