--- packages/core/src/components/Edges/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './general'
-export * from './simple-bezier'
